// src/modules/organizations/domain/enums/organization-type.ts
export enum EnterpriseType {
    ACADEMIA = 'ACADEMIA',
    EMPRESA_INTERNACIONAL = 'EMPRESA_INTERNACIONAL',
    EMPRESA_NACIONAL = 'EMPRESA_NACIONAL',
    GOBIERNO_NACIONAL = 'GOBIERNO_NACIONAL',
    INDEPENDIENTE = 'INDEPENDIENTE',
    ONG = 'ONG',
    ORGANISMO_INTERNACIONAL = 'ORGANISMO_INTERNACIONAL',
}